code working when split in two components
